+++ src/main/java/com/alex/spring/vendas/repositories/ClientRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+public interface ClientRepository extends JpaRepository<Client, Long> {
     Page<Client> findByNameIgnoreCase(String name, Pageable pageable);
 
     Page<Client> findByGender(Gender gender, Pageable pageable);
