+++ src/main/java/com/alex/spring/vendas/repositories/SaleRepository.java
@@ -3,5 +3,5 @@ package com.alex.spring.vendas.repositories;
 import com.alex.spring.vendas.domain.Sale;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SaleRepository extends JpaRepository<Sale, Integer> {
+public interface SaleRepository extends JpaRepository<Sale, Long> {
 }
